replace schedulers & add test case for base classes
