+++ zae_engine/schedulers/core.py
@@ -1,4 +1,3 @@
-import math
 from abc import ABC, abstractmethod
 
 from torch.optim import lr_scheduler, Optimizer
@@ -56,25 +55,3 @@ class SchedulerChain(SchedulerBase):
 
     def get_lr(self):
         return self.schedulers[self.i_scheduler].get_lr()
-
-
-class WarmUpScheduler(SchedulerBase):
-    def __init__(self, optimizer: Optimizer, total_steps, eta_min, last_epoch: int = -1):
-        super(WarmUpScheduler, self).__init__(optimizer, total_steps, eta_min, last_epoch)
-
-    def get_lr(self):
-        return [self.eta_min + (lr - self.eta_min) * self._step_count / self.total_steps for lr in self.base_lrs]
-
-
-class CosineAnnealingScheduler(SchedulerBase):
-    def __init__(self, optimizer: Optimizer, total_steps, eta_min, last_epoch: int = -1):
-        super(CosineAnnealingScheduler, self).__init__(optimizer, total_steps, eta_min, last_epoch)
-
-    def get_lr(self):
-        lrs = []
-        for lr in self.base_lrs:
-            radian = math.pi * self._step_count / self.total_steps
-            annealing = 0.5 * (math.cos(radian) + 1)
-            offset = self.eta_min
-            lrs.append(offset + (lr - offset) * annealing)
-        return lrs
